+++ backend/auth_service/auth_app/schemas/company.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from pydantic import BaseModel, Field, ConfigDict
 from uuid import UUID
 from typing import List, Optional
@@ -7,7 +8,7 @@ class CompanyBase(BaseModel):
     name: str = Field(..., min_length=1, max_length=100)
     logo: Optional[str] = None # Added logo field
     base_currency: str = Field("USD", min_length=3, max_length=3)
-    default_tax_rate: float = 0.16
+    default_tax_rate: Decimal = 0.16
 
 # Properties to receive on company creation
 class CompanyCreate(CompanyBase):
