+++ backend/inventory_service/inventory_app/schemas/pos.py
@@ -7,7 +7,7 @@ from common.domain.value_objects import Money
 class SaleItemCreate(BaseModel):
     product_id: uuid.UUID
     sku: str
-    quantity: float = Field(..., gt=0)
+    quantity: Decimal = Field(..., gt=0)
     unit_price: Decimal = Field(..., ge=Decimal("0.0"))
     currency: str = "MXN"
     is_taxable: bool = True
@@ -17,7 +17,7 @@ class SaleCreate(BaseModel):
     warehouse_id: uuid.UUID
     comments: Optional[str] = None
     customer_id: Optional[uuid.UUID] = None # Optional for POS
-    total_amount: float
+    total_amount: Decimal
     currency: str = "MXN"
 
 class SaleResponse(BaseModel):
