+++ backend/mes_service/mes_app/models/run_metrics_snapshot.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 import uuid
 from sqlalchemy import ForeignKey, Numeric
 from sqlalchemy.orm import Mapped, mapped_column
@@ -9,10 +10,10 @@ class RunMetricsSnapshot(MultiTenantBase):
     
     production_run_id: Mapped[uuid.UUID] = mapped_column(ForeignKey("mes_production_runs.id"), unique=True, nullable=False)
     
-    availability: Mapped[float] = mapped_column(Numeric(5, 4), default=0.0) # E.g. 0.95 (95%)
-    efficiency: Mapped[float] = mapped_column(Numeric(5, 4), default=0.0)
-    quality: Mapped[float] = mapped_column(Numeric(5, 4), default=0.0)
-    oee: Mapped[float] = mapped_column(Numeric(5, 4), default=0.0)
+    availability: Mapped[Decimal] = mapped_column(Numeric(5, 4), default=0.0) # E.g. 0.95 (95%)
+    efficiency: Mapped[Decimal] = mapped_column(Numeric(5, 4), default=0.0)
+    quality: Mapped[Decimal] = mapped_column(Numeric(5, 4), default=0.0)
+    oee: Mapped[Decimal] = mapped_column(Numeric(5, 4), default=0.0)
     
-    tak_time_seconds: Mapped[float] = mapped_column(Numeric(10, 4), default=0.0)
-    lmpu_minutes: Mapped[float] = mapped_column(Numeric(10, 4), default=0.0) # Labor Minutes Per Unit
+    tak_time_seconds: Mapped[Decimal] = mapped_column(Numeric(10, 4), default=0.0)
+    lmpu_minutes: Mapped[Decimal] = mapped_column(Numeric(10, 4), default=0.0) # Labor Minutes Per Unit
