+++ backend/wms_service/wms_app/models/sales_order.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 import enum
 import uuid
 from typing import Optional
@@ -27,7 +28,7 @@ class SalesOrder(MultiTenantBase):
     product_id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), index=True, nullable=True)
     warehouse_id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), index=True, nullable=True)
     uom_id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), nullable=True)
-    quantity: Mapped[float] = mapped_column(Numeric(15, 4), default=0.0)
+    quantity: Mapped[Decimal] = mapped_column(Numeric(15, 4), default=0.0)
     
     total_items: Mapped[int] = mapped_column(Integer, default=0, nullable=False)
     observations: Mapped[Optional[str]] = mapped_column(String(500), nullable=True)
